move settings around
