+++ src/entries/popup/pages/settings/walletsAndKeys/privateKey/warning.tsx
@@ -6,7 +6,7 @@ import { IconAndCopyItem } from '~/entries/popup/components/IconAndCopyList.tsx/
 import WarningInfo from '~/entries/popup/components/WarningInfo/WarningInfo';
 import { ROUTES } from '~/entries/popup/urls';
 
-import { ConfirmPasswordPrompt } from '../../confirmPasswordPrompt';
+import { ConfirmPasswordPrompt } from '../../privacy/confirmPasswordPrompt';
 
 const iconAndCopyList: IconAndCopyItem[] = [
   {
@@ -15,7 +15,7 @@ const iconAndCopyList: IconAndCopyItem[] = [
       color: 'orange',
     },
     copy: i18n.t(
-      'settings.privacy_and_security.wallets_and_keys.recovery_phrase.warning_1',
+      'settings.privacy_and_security.wallets_and_keys.private_key.warning_1',
     ),
   },
   {
@@ -24,7 +24,7 @@ const iconAndCopyList: IconAndCopyItem[] = [
       color: 'pink',
     },
     copy: i18n.t(
-      'settings.privacy_and_security.wallets_and_keys.recovery_phrase.warning_2',
+      'settings.privacy_and_security.wallets_and_keys.private_key.warning_2',
     ),
   },
   {
@@ -33,7 +33,7 @@ const iconAndCopyList: IconAndCopyItem[] = [
       color: 'red',
     },
     copy: i18n.t(
-      'settings.privacy_and_security.wallets_and_keys.recovery_phrase.warning_3',
+      'settings.privacy_and_security.wallets_and_keys.private_key.warning_3',
     ),
   },
   {
@@ -42,12 +42,11 @@ const iconAndCopyList: IconAndCopyItem[] = [
       color: 'blue',
     },
     copy: i18n.t(
-      'settings.privacy_and_security.wallets_and_keys.recovery_phrase.warning_4',
+      'settings.privacy_and_security.wallets_and_keys.private_key.warning_4',
     ),
   },
 ];
-
-export function RecoveryPhraseWarning() {
+export function PrivateKeyWarning() {
   const { state } = useLocation();
   const [showEnterPassword, setShowEnterPassword] = useState(false);
 
@@ -64,18 +63,17 @@ export function RecoveryPhraseWarning() {
         show={showEnterPassword}
         onClose={closePasswordPrompt}
         redirect={
-          ROUTES.SETTINGS__PRIVACY__WALLETS_AND_KEYS__WALLET_DETAILS__RECOVERY_PHRASE
+          ROUTES.SETTINGS__PRIVACY__WALLETS_AND_KEYS__WALLET_DETAILS__PKEY
         }
         extraState={{ ...state }}
       />
       <WarningInfo
-        testId={'show-phrase'}
-        iconAndCopyList={iconAndCopyList}
         onProceed={openPasswordPrompt}
+        iconAndCopyList={iconAndCopyList}
         proceedButtonLabel={i18n.t(
-          'settings.privacy_and_security.wallets_and_keys.recovery_phrase.show',
+          'settings.privacy_and_security.wallets_and_keys.private_key.show',
         )}
-        proceedButtonSymbol="doc.plaintext.fill"
+        proceedButtonSymbol="key.fill"
       />
     </>
   );
