+++ src/entries/popup/pages/settings/walletsAndKeys/recoveryPhrase/recoveryPhrase.tsx
