+++ src/entries/popup/pages/settings/walletsAndKeys/recoveryPhrase/recoveryPhraseVerify.tsx
