+++ src/entries/popup/pages/settings/walletsAndKeys/wipeWalletWarning.tsx
@@ -8,8 +8,8 @@ import { IconAndCopyItem } from '~/entries/popup/components/IconAndCopyList.tsx/
 import WarningInfo from '~/entries/popup/components/WarningInfo/WarningInfo';
 import { ROUTES } from '~/entries/popup/urls';
 
-import * as wallet from '../../../../handlers/wallet';
-import { ConfirmPasswordPrompt } from '../confirmPasswordPrompt';
+import * as wallet from '../../../handlers/wallet';
+import { ConfirmPasswordPrompt } from '../privacy/confirmPasswordPrompt';
 
 const iconAndCopyList: IconAndCopyItem[] = [
   {
